+++ tests/integration/Migration/Action/GenerateCest.php
@@ -18,6 +18,7 @@ use Phalcon\Db\Column;
 use Phalcon\Db\Reference;
 use Phalcon\Migrations\Exception\Db\UnknownColumnTypeException;
 use Phalcon\Migrations\Migration\Action\Generate;
+use Phalcon\Migrations\Mvc\Model\Migration;
 
 final class GenerateCest
 {
@@ -215,5 +216,82 @@ final class GenerateCest
         $I->assertSame(count($columns), count($class->getQuoteWrappedColumns()));
         $I->assertSame("'column1'", $class->getQuoteWrappedColumns()[0]);
         $I->assertSame("'column2'", $class->getQuoteWrappedColumns()[1]);
+    }
+
+    /**
+     * @param IntegrationTester $I
+     */
+    public function throwUnknownColumnTypeException(IntegrationTester $I): void
+    {
+        $I->wantToTest('Migration\Action\Generate - getColumns() throw UnknownColumnTypeException');
+
+        $I->expectThrowable(UnknownColumnTypeException::class, function () {
+            $columns = [
+                new Column('unknown', [
+                    'type' => 9000,
+                    'size' => 10,
+                    'notNull' => true,
+                ]),
+            ];
+
+            $data = [];
+            $class = new Generate('mysql', $columns);
+            foreach ($class->getColumns() as $column) {
+                // Wait error
+                $data[] = $column;
+            }
+        });
+    }
+
+    /**
+     * @param IntegrationTester $I
+     * @throws UnknownColumnTypeException
+     */
+    public function columnHasDefault(IntegrationTester $I): void
+    {
+        $I->wantToTest('Migration\Action\Generate - getColumns() has default');
+
+        $expected = "'default' => \"0\"";
+        $columnsWithDefault = [
+            new Column('column_default', [
+                'type' => Column::TYPE_INTEGER,
+                'size' => 10,
+                'notNull' => true,
+                'default' => 0,
+            ]),
+        ];
+
+        $columnsWithDefaultAndAI = [
+            new Column('column_ai', [
+                'type' => Column::TYPE_INTEGER,
+                'size' => 10,
+                'notNull' => true,
+                'autoIncrement' => true,
+                'first' => true,
+                'primary' => true,
+                'default' => 0,
+            ]),
+        ];
+
+        $class1 = new Generate(Migration::DB_ADAPTER_MYSQL, $columnsWithDefault);
+        $class2 = new Generate(Migration::DB_ADAPTER_MYSQL, $columnsWithDefaultAndAI);
+        $class3 = new Generate(Migration::DB_ADAPTER_POSTGRESQL, $columnsWithDefault);
+        $class4 = new Generate(Migration::DB_ADAPTER_POSTGRESQL, $columnsWithDefaultAndAI);
+        $class5 = new Generate(Migration::DB_ADAPTER_SQLITE, $columnsWithDefault);
+        $class6 = new Generate(Migration::DB_ADAPTER_SQLITE, $columnsWithDefaultAndAI);
+
+        $array1 = current(iterator_to_array($class1->getColumns()));
+        $array2 = current(iterator_to_array($class2->getColumns()));
+        $array3 = current(iterator_to_array($class3->getColumns()));
+        $array4 = current(iterator_to_array($class4->getColumns()));
+        $array5 = current(iterator_to_array($class5->getColumns()));
+        $array6 = current(iterator_to_array($class6->getColumns()));
+
+        $I->assertSame($expected, $array1[1]);
+        $I->assertFalse(in_array($expected, $array2));
+        $I->assertSame($expected, $array3[1]);
+        $I->assertFalse(in_array($expected, $array4));
+        $I->assertSame($expected, $array5[1]);
+        $I->assertFalse(in_array($expected, $array6));
     }
 }